Fix broken links to meeting minutes
Add all the minuts to SUMMARY.md, else they won't be included in the
site, even though they're linked to from other pages.

This is a (dubious) requirement  of mdbook - only pages linked from the
SUMMARY.md are rendered.

I guess it's more reasonable given the primary focus of mdbook is to
render a book-like document, not a wiki, so why would you have a chapter
of your book that's not navigable to from the TOC?

Anyway, there is an outstanding feature requests for this kind of
behavior: https://github.com/rust-lang/mdBook/issues/702
No one has implemented it for a long time - but the maintainer doesn't
seem opposed to the feature in principal.

We have a few ways forward:

1. live with it for now and hope someone eventually adds this
   functionality
2. port the wiki to some other platform. I'm not that attached to
   mdbook, it's just simple and familiar for me.